requirements.txt
langchain
langchain-openai
langgraph
openai
python-dotenv
pydantic
faiss-cpu
tiktoken
langchain-community